--- src/ndarray/operations.rs
@@ -1,258 +1,0 @@
-use crate::ndarray::NdArrayFlags;
-use crate::broadcast::broadcast_shapes;
-use crate::{IntegerDataType, RawDataType, NdArray};
-use std::ops::{Add, AddAssign, BitAnd, BitAndAssign, BitOr, BitOrAssign, Div, DivAssign, Mul, MulAssign, Neg, Rem, RemAssign, Shl, ShlAssign, Shr, ShrAssign, Sub, SubAssign};
-
-use paste::paste;
-use crate::backwards::{AddBackwards, DivBackwards, MulBackwards, NegBackwards, SubBackwards};
-
-macro_rules! define_binary_ops {
-    ($($trait_: ident, $operator: tt, $method: ident;)* ) => {
-        $(
-            fn $method<'a, 'b>(lhs: impl AsRef<NdArray<'a, T>>,
-                               rhs: impl AsRef<NdArray<'b, T>>) -> NdArray<'static, T>
-            where
-                T: $trait_<Output=T>,
-            {
-                let lhs = lhs.as_ref();
-                let rhs = rhs.as_ref();
-
-                let shape = broadcast_shapes(&lhs.shape, &rhs.shape);
-                let lhs = lhs.broadcast_to(&shape);
-                let rhs = rhs.broadcast_to(&shape);
-
-                let data = lhs.flatiter().zip(rhs.flatiter()).map(|(lhs, rhs)| lhs $operator rhs).collect();
-                unsafe { NdArray::from_contiguous_owned_buffer(shape, data, false, false) }
-            }
-
-            paste! { fn [<$method _scalar>] <'a, 'b>(lhs: impl AsRef<NdArray<'a, T>>,
-                                                     rhs: T) -> NdArray<'static, T>
-                where
-                    T: $trait_<Output=T>,
-                {
-                    let lhs = lhs.as_ref();
-
-                    let data = lhs.flatiter().map(|lhs| lhs $operator rhs).collect();
-                    unsafe { NdArray::from_contiguous_owned_buffer(lhs.shape.clone(), data, false, false) }
-                }
-            }
-        )*
-    }
-}
-
-macro_rules! define_binary_ops_with_grad {
-    ($dtype:ty, $($trait_: ident, $operator: tt, $method: ident, $backwards: ident;)* ) => {
-        $(
-            fn $method<'a, 'b>(lhs: impl AsRef<NdArray<'a, $dtype>>,
-                               rhs: impl AsRef<NdArray<'b, $dtype>>) -> NdArray<'static, $dtype>
-            {
-                let lhs = lhs.as_ref();
-                let rhs = rhs.as_ref();
-
-                let requires_grad = lhs.requires_grad() || rhs.requires_grad();
-
-                let shape = broadcast_shapes(&lhs.shape, &rhs.shape);
-                let broadcasted_lhs = lhs.broadcast_to(&shape);
-                let broadcasted_rhs = rhs.broadcast_to(&shape);
-
-                let data = broadcasted_lhs.flatiter()
-                                        .zip(broadcasted_rhs.flatiter())
-                                        .map(|(a, b)| a $operator b)
-                                        .collect();
-
-                let mut result = unsafe { NdArray::from_contiguous_owned_buffer(shape, data, requires_grad, false) };
-
-                if requires_grad {
-                    result.grad_fn = $backwards::new(lhs, rhs);
-                }
-
-                result
-            }
-
-            paste! { fn [<$method _scalar>] <'a, 'b>(lhs: impl AsRef<NdArray<'a, $dtype>>,
-                                                     rhs: $dtype) -> NdArray<'static, $dtype>
-                {
-                    let lhs = lhs.as_ref();
-
-                    let requires_grad = lhs.requires_grad();
-
-                    let data = lhs.flatiter().map(|lhs| lhs $operator rhs).collect();
-                    unsafe { NdArray::from_contiguous_owned_buffer(lhs.shape.clone(), data, requires_grad, false) }
-                }
-            }
-        )*
-
-        fn neg<'a, 'b>(rhs: impl AsRef<NdArray<'a, $dtype>>) -> NdArray<'static, $dtype> {
-            let rhs = rhs.as_ref();
-
-            let requires_grad = rhs.requires_grad();
-
-            let data = rhs.flatiter().map(|rhs| -rhs).collect();
-            let mut result = unsafe { NdArray::from_contiguous_owned_buffer(rhs.shape.clone(), data, requires_grad, false) };
-
-            if requires_grad {
-                result.grad_fn = NegBackwards::new(rhs);
-            }
-
-            result
-        }
-    }
-}
-
-macro_rules! implement_binary_ops {
-    ($($trait_: ident, $method: ident;)* ) => {
-        impl<T: RawDataType + Neg<Output=T>> Neg for NdArray<'_, T> {
-            type Output = NdArray<'static, T>;
-        
-            fn neg(self) -> Self::Output { <T as TensorBinaryOps<T>>::neg(self) }
-        }
-        
-        impl<T: RawDataType + Neg<Output=T>> Neg for &NdArray<'_, T> {
-            type Output = NdArray<'static, T>;
-        
-            fn neg(self) -> Self::Output { <T as TensorBinaryOps<T>>::neg(self) }
-        }
-                
-        $(
-            impl<T: RawDataType + $trait_<Output=T>> $trait_ for NdArray<'_, T> {
-                type Output = NdArray<'static, T>;
-
-                fn $method(self, rhs: NdArray<T>) -> Self::Output { <T as TensorBinaryOps<T>>::$method(self, rhs) }
-            }
-            impl<T: RawDataType + $trait_<Output=T>> $trait_<&NdArray<'_, T>> for NdArray<'_, T> {
-                type Output = NdArray<'static, T>;
-
-                fn $method(self, rhs: &NdArray<T>) -> Self::Output { <T as TensorBinaryOps<T>>::$method(self, rhs) }
-            }
-            impl<T: RawDataType + $trait_<Output=T>> $trait_<NdArray<'_, T>> for &NdArray<'_, T> {
-                type Output = NdArray<'static, T>;
-
-                fn $method(self, rhs: NdArray<T>) -> Self::Output { <T as TensorBinaryOps<T>>::$method(self, rhs) }
-            }
-            impl<T: RawDataType + $trait_<Output=T>> $trait_<&NdArray<'_, T>> for &NdArray<'_, T> {
-                type Output = NdArray<'static, T>;
-
-                fn $method(self, rhs: &NdArray<T>) -> Self::Output { <T as TensorBinaryOps<T>>::$method(self, rhs) }
-            }
-
-            impl<T: RawDataType + $trait_<Output=T>> $trait_<T> for NdArray<'_, T> {
-                type Output = NdArray<'static, T>;
-
-                fn $method(self, rhs: T) -> Self::Output { paste! { <T as TensorBinaryOps<T>>::[<$method _scalar>](self, rhs) } }
-            }
-            impl<T: RawDataType + $trait_<Output=T>> $trait_<T> for &NdArray<'_, T> {
-                type Output = NdArray<'static, T>;
-
-                fn $method(self, rhs: T) -> Self::Output { paste! { <T as TensorBinaryOps<T>>::[<$method _scalar>](self, rhs) } }
-            }
-        )*
-    };
-
-    ($dtype1:ty, $dtype2:ty, $($trait_: ident, $method: ident;)* ) => {
-        implement_binary_ops!($dtype1, $($trait_, $method;)* );
-        implement_binary_ops!($dtype2, $($trait_, $method;)* );
-    };
-
-    ($dtype1:ty, $dtype2:ty, $dtype3:ty, $dtype4:ty, $($trait_: ident, $method: ident;)* ) => {
-        implement_binary_ops!($dtype1, $dtype2, $($trait_, $method;)* );
-        implement_binary_ops!($dtype3, $dtype4, $($trait_, $method;)* );
-        implement_binary_ops!($dtype5, $dtype6, $($trait_, $method;)* );
-    }
-}
-
-pub trait TensorBinaryOps<T: RawDataType> {
-    define_binary_ops!(
-        Add, +, add;
-        Sub, -, sub;
-        Mul, *, mul;
-        Div, /, div;
-        Rem, %, rem;
-        BitAnd, &, bitand;
-        BitOr, |, bitor;
-        Shl, <<, shl;
-        Shr, >>, shr;
-    );
-
-    fn neg<'a, 'b>(rhs: impl AsRef<NdArray<'a, T>>) -> NdArray<'static, T>
-    where
-        T: Neg<Output=T>,
-    {
-        let rhs = rhs.as_ref();
-
-        let data = rhs.flatiter().map(|rhs| -rhs).collect();
-        unsafe { NdArray::from_contiguous_owned_buffer(rhs.shape.clone(), data, false, false) }
-    }
-}
-
-implement_binary_ops!(
-    Add, add;
-    Sub, sub;
-    Mul, mul;
-    Div, div;
-    Rem, rem;
-    BitAnd, bitand;
-    BitOr, bitor;
-    Shl, shl;
-    Shr, shr;
-);
-
-impl<T: IntegerDataType> TensorBinaryOps<T> for T {}
-impl TensorBinaryOps<bool> for bool {}
-
-impl TensorBinaryOps<f32> for f32 {
-    define_binary_ops_with_grad!(
-        f32,
-        Add, +, add, AddBackwards;
-        Sub, -, sub, SubBackwards;
-        Mul, *, mul, MulBackwards;
-        Div, /, div, DivBackwards;
-    );
-}
-
-impl TensorBinaryOps<f64> for f64 {
-    define_binary_ops_with_grad!(
-        f64,
-        Add, +, add, AddBackwards;
-        Sub, -, sub, SubBackwards;
-        Mul, *, mul, MulBackwards;
-        Div, /, div, DivBackwards;
-    );
-}
-
-macro_rules! define_binary_iop {
-    ( $trait_: ident, $operator: tt, $method: ident ) => {
-    impl<T: RawDataType + $trait_> $trait_<NdArray<'_, T>> for NdArray<'_, T> {
-        fn $method(&mut self, rhs: NdArray<'_, T>) {
-            *self $operator &rhs
-        }
-    }
-
-    impl<T: RawDataType + $trait_> $trait_<&NdArray<'_, T>> for NdArray<'_, T> {
-        fn $method(&mut self, rhs: &NdArray<'_, T>) {
-            if !self.flags.contains(NdArrayFlags::Writeable) {
-                panic!("tensor is readonly.");
-            }
-
-            if self.requires_grad() && self.is_leaf() {
-                panic!("a leaf tensor that requires grad cannot be used in an in-place operation.");
-            }
-
-            let rhs = rhs.broadcast_to(&self.shape);
-
-            for (lhs, rhs) in self.flatiter_ptr().zip(rhs.flatiter()) {
-                unsafe { *lhs $operator rhs; }
-            }
-        }
-    }
-    };
-}
-
-define_binary_iop!(AddAssign, +=, add_assign);
-define_binary_iop!(SubAssign, -=, sub_assign);
-define_binary_iop!(MulAssign, *=, mul_assign);
-define_binary_iop!(DivAssign, /=, div_assign);
-define_binary_iop!(RemAssign, %=, rem_assign);
-define_binary_iop!(BitAndAssign, &=, bitand_assign);
-define_binary_iop!(BitOrAssign, |=, bitor_assign);
-define_binary_iop!(ShlAssign, <<=, shl_assign);
-define_binary_iop!(ShrAssign, >>=, shr_assign);
