+++ src/util/dtype.rs
@@ -1,4 +1,4 @@
-use crate::operations::TensorBinaryOps;
+use crate::binary_ops::BinaryOps;
 use num::traits::MulAdd;
 use num::{Bounded, Float, NumCast, One, ToPrimitive, Zero};
 use rand::distributions::uniform::SampleUniform;
@@ -6,7 +6,7 @@ use std::fmt::{Debug, Display};
 use std::iter::{Product, Sum};
 use std::ops::{Add, AddAssign, Div, Mul, MulAssign, Neg, Sub, SubAssign};
 
-pub trait RawDataType: Clone + Copy + PartialEq + Display + Default + Debug + Send + Sync + TensorBinaryOps<Self> + 'static {}
+pub trait RawDataType: Clone + Copy + PartialEq + Display + Default + Debug + Send + Sync + BinaryOps<Self> + 'static {}
 
 impl RawDataType for u8 {}
 impl RawDataType for u16 {}
